refactor(projection): extract curated layer load and Supabase reload
